requirements.txt
requests
python-dotenv
playwright
pyautogui